[DSP-116] Make FA1.2 forwarder accept an arbitrary contract address
Problem: FA1.2 specialized forwarder accepts ManagedLedger contract
reference. Since we want to use FA1.2 forwarder not only with ML
but with other contracts (with different parameters) as well, we
need the forwarder to accept an arbitrary contract address.

Solution: Change (ContractRef ML.Parameter) to just Address. Use
the entrypoints feature to call the stored contract. Fail at
runtime if the supplied address is not a valid FA1.2 contract.