[DSP-116] Add network test suite
Problem: Although tests written in Lorentz eDSL give some
confidence in the code, the real network can pose new challenges
(like the possibility of gas exhaustion) that can't be caught by
the eDSL tests.

Solution: Add a test suite based on `morley-nettest` package that
executes the given scenario on a real network. Currently, the
scenario lacks some important test cases (like the most expensive
transfers), these are to be added in a separate commit. The test
suite does not yet allow to choose between different flavors of
the forwarder contract – this would be also added later.